Add files via upload
Disclaimer: None of the libraries used are owned or created by me. This is simply a demonstration of making a functional(and albeit quite inefficient) keyboard using the minimum number of i/o pins while still being useful. The above program is used to connect to a wifi network to send whatever prompts you might like.